Refactor: Domain에 plan -> travelplan entity패키지삭제
