+++ src/main/java/grepp/NBE5_6_2_Team03/api/controller/travelplan/TravelPlanController.java
@@ -1,8 +1,8 @@
 package grepp.NBE5_6_2_Team03.api.controller.travelplan;
 
 import grepp.NBE5_6_2_Team03.api.controller.travelplan.dto.TravelPlanRequestDto;
-import grepp.NBE5_6_2_Team03.domain.plan.entity.TravelPlan;
-import grepp.NBE5_6_2_Team03.domain.plan.service.TravelPlanService;
+import grepp.NBE5_6_2_Team03.domain.travelplan.TravelPlan;
+import grepp.NBE5_6_2_Team03.domain.travelplan.service.TravelPlanService;
 import grepp.NBE5_6_2_Team03.domain.user.CustomUserDetails;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
