+++ src/main/java/grepp/NBE5_6_2_Team03/domain/schedule/TravelSchedule.java
@@ -1,6 +1,6 @@
 package grepp.NBE5_6_2_Team03.domain.schedule;
 
-import grepp.NBE5_6_2_Team03.domain.plan.entity.TravelPlan;
+import grepp.NBE5_6_2_Team03.domain.travelplan.TravelPlan;
 import grepp.NBE5_6_2_Team03.domain.schedule.code.ScheduleStatus;
 import jakarta.persistence.*;
 import lombok.*;
