+++ src/main/java/grepp/NBE5_6_2_Team03/domain/travelplan/repository/TravelPlanRepository.java
@@ -1,6 +1,6 @@
-package grepp.NBE5_6_2_Team03.domain.plan.repository;
+package grepp.NBE5_6_2_Team03.domain.travelplan.repository;
 
-import grepp.NBE5_6_2_Team03.domain.plan.entity.TravelPlan;
+import grepp.NBE5_6_2_Team03.domain.travelplan.TravelPlan;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
